Update asc() to call setup-ssh-screen if it exists, else screen as usual
With the exception of ssh agent-forwarding now being enabled by
default this works as before if no ~/.mutt/bin/setup-ssh-screen script
exists on the remote site.

If that script exists, it will take care of bringing the ssh-agent
to work inside the reattached screen session.

See http://www.stderr.nl/Blog/Software/Mutt/MuttRemoteAttachments.html
for the script and details and
http://samrowe.com/wordpress/ssh-agent-and-gnu-screen/ for the initial
idea.