Create a function md which creates a directory and changes to it
As grml removed the md alias (for "mkdir -p") and renamed the function
mcd (which did the same as this function) to a longer mkcd, I take the
chance to make it the way I always wanted it.

Just to be sure I still check if an md alias exists (which would
overrule the function) and remove it in that case.