Reintroduce agrep as what I currently expect it to do
Actually I always think agrep is part of atool like adiff. But atool
doesn't provide an agrep. patool though does, but under a different
name. So let's introduce agrep as alias to "patool search" and ignore
the fact that /usr/bin/agrep exists, but is not what I expect it to
do. (And it's just in non-free anyways.)